Move sync implementation into entity-specific modules to make sync.ts cleaner
